+++ CHANGELOG.md
@@ -5,6 +5,8 @@ changes to any options/interfaces the checker exposes for developers.
 # 17.N.N
 NN XXX 2017
   - Allow `&;` (don’t report it as “`&` did not start a character reference”)
+  - Update CSP checking to Salvation 2.3.0
+    https://github.com/shapesecurity/salvation/releases/tag/v2.3.0
 
 # 17.7.0
 26 June 2017